+++ src/styles/theme.tsx
@@ -1,4 +1,5 @@
 export const Theme = {
+  complement: '#B0791E',
   gray: '#E7E7E7',
   primary: '#254463',
   secondary: '#DFEFFF',