+++ docs/changelog.md
@@ -33,6 +33,7 @@ Sprint 12 / May 10, 2020
 - updated readme.
 - added App component to documentation.
 - added ability to log out using React Router's Prompt.
+- added `complement` color to theme.
 
 <hr />